Update modelzoo Signed-off-by: weijingchen <talkingwallace@sohu.com>
